Fix the commit mess

+++ include/utils.h
@@ -29,12 +29,12 @@ char *getLine(char *);
 
 void moduleAddr(pid_t pid, char *lib, int32_t *start, int32_t *end);
 
-char pokeAddr(pid_t, long, char *, int);
+char pokeAddr(pid_t, int32_t, char *, size_t);
 
 char readAddr(pid_t, unsigned int, void *, size_t);
 
 off_t ScanAddr(int32_t, int32_t, const char *);
 
-char writeAddr(pid_t, ssize_t, void *, ssize_t);
+char writeAddr(pid_t, int32_t, void *, size_t);
 
 #endif /* _UTILS_H */
